break up dependency on formatter
